# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest-Api for library",
        "title": "BibliotecaAPI",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/api/admin/reservations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get active reservations with filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas"
                ],
                "summary": "Get active reservations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "idsocio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "idlibro",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reservation date",
                        "name": "fechareserva",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User name",
                        "name": "nombre_socio",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Reservation"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get users filtered by type (normal, estudiante, profesor)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Get users by type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User type (normal, estudiante, profesor)",
                        "name": "tiposocio",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.User_"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/users/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all users in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.User_"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/reservation": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new book reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas"
                ],
                "summary": "Create reservation",
                "parameters": [
                    {
                        "description": "Reservation data",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Reservation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.Reservation"
                        }
                    }
                }
            }
        },
        "/api/reservations": {
            "get": {
                "description": "Get active reservations for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas"
                ],
                "summary": "Get user active reservations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "usuario_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Reservation"
                            }
                        }
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel an existing reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservas"
                ],
                "summary": "Cancel reservation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation cancelled successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Reservation": {
            "type": "object",
            "properties": {
                "fechareserva": {
                    "type": "string"
                },
                "idlibro": {
                    "type": "integer"
                },
                "idsocio": {
                    "type": "integer"
                }
            }
        },
        "data.User_": {
            "type": "object",
            "properties": {
                "correo": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },
                "fechanacimiento": {
                    "type": "string"
                },
                "fecharegistro": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "rol": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                },
                "tiposocio": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}